refactor: Rename `…lugins/aws/package/compile/events/cognitoUserPool.js`
