+++ lib/plugins/aws/package/compile/events/cognito-user-pool.js
