+++ lib/plugins/index.js
@@ -48,7 +48,7 @@ module.exports = [
   require('./aws/package/compile/events/iotFleetProvisioning.js'),
   require('./aws/package/compile/events/cloud-watch-event.js'),
   require('./aws/package/compile/events/cloud-watch-log.js'),
-  require('./aws/package/compile/events/cognitoUserPool.js'),
+  require('./aws/package/compile/events/cognito-user-pool.js'),
   require('./aws/package/compile/events/eventBridge/index.js'),
   require('./aws/package/compile/events/sqs.js'),
   require('./aws/package/compile/events/cloud-front.js'),
